Refactor to one file
